+++ webui/src/App.tsx
@@ -3,6 +3,7 @@ import { useRuntimeState } from './api/hooks'
 import { AppShell } from './components/layout/AppShell'
 import { NeuralNetworkPage } from './pages/NeuralNetworkPage'
 import { OperationsPage } from './pages/OperationsPage'
+import { OptionsStrategyPage } from './pages/OptionsStrategyPage'
 import { Overview } from './pages/Overview'
 import { RiskPage } from './pages/RiskPage'
 import { TopologyPage } from './pages/TopologyPage'
@@ -17,6 +18,7 @@ function App() {
         <Route path="/" element={<Overview state={state} />} />
         <Route path="/operations" element={<OperationsPage state={state} />} />
         <Route path="/risk" element={<RiskPage state={state} />} />
+        <Route path="/options-strategy" element={<OptionsStrategyPage state={state} />} />
         <Route path="/topology" element={<TopologyPage state={state} />} />
         <Route path="/neural-network" element={<NeuralNetworkPage state={state} />} />
         <Route path="/tracing" element={<TracingPage />} />